re-image and did some request
some of the guns now have normal casings and now empty a mag or something els will now give you the right bullet for it. add in that canister image hase replace the phernome one.   add in new image calld grave. now you can have empty virsion of drinks to use in mapping. add a texture for all the ammo bullets and casings.