Added mines, comissary, jobs to Camp
Several addictions to POW camp, with roles for prisoners. Added mines
with mineable ore, and a comissary for prisoners to buy stuff.

Also, there are now armbands assigned to prisoner jobs: Red and white
for collaborator security, white for medical personel, green and blue
for kitchen duty, yellow for miners, purple for janitors.