Adds shemagh, and back layer fix
Adds shemagh facewear which have 4 states that can be switched to. Can be set to cover the neck only, neck and mouth, neck mouth and head, neck and head.

Also a fix so that items in the back slot don't get hidden by headwear and facewear, especially when facing the north.